Move base and vector_parameter_study to subpackage
The **methods** subpackage will hold the implementations of Dakota
methods.
