+++ dakota/__init__.py
@@ -1,0 +1,3 @@
+"""Interface to the Dakota iterative systems toolkit."""
+
+method_path = 'dakota.methods.'
