+++ dakota/methods/__init__.py
@@ -1,0 +1,7 @@
+"""Iterative methods provided by Dakota.
+
+The module name must match the keyword used by Dakota for the method;
+e.g., the Dakota keyword ``vector_parameter_study`` is used to name
+**vector_parameter_study.py**.
+
+"""
