+++ dakota/tests/test_vector_parameter_study.py
@@ -8,7 +8,7 @@
 # Mark Piper (mark.piper@colorado.edu)
 
 from nose.tools import *
-from dakota.vector_parameter_study import VectorParameterStudy, method
+from dakota.methods.vector_parameter_study import VectorParameterStudy, method
 from . import start_dir, data_dir
